Overlay render target, sorting & blending
This commit introduces render targets and implements two of them: the
framebuffer (where normal scene rendering goes) and overlay, which is rendered
on top of the framebuffer.

The render queue is now sorted based on various criteria, the most important is
Z-based sorting for text and quad render operations, in order to implement
proper blending and translucency.